Cast a test-only quartc constant to int64_t so it can be shifted by 32 on android without overflowing.
PiperOrigin-RevId: 245968333

+++ quic/quartc/test/random_delay_link.cc
@@ -17,7 +17,7 @@ namespace simulator {
 namespace {
 
 // Number of buckets used to define an exponential distribution.
-constexpr int64_t kNumBuckets = 2l << 32;
+constexpr int64_t kNumBuckets = static_cast<int64_t>(2) << 32;
 
 }  // namespace
